Update colorAnalysisWorker.ts to add more CommonScales for pixel art detection and reduction
In the pixel art detection code, used when trying to automatically detect a scaled pixel art image and unscale it to its original resolution, CommonScales now are from 2 to 70 without missing numbers in the middle. 

This should help to better detect if an image is an scaled pixel art, reducing the false negatives.